+++ docker-compose.local.yml
@@ -34,21 +34,21 @@ services:
     volumes:
       - redis_data:/data
 
-  elasticsearch:
-    image: elasticsearch:7.17.5
-    restart: always
-    ports:
-      - "9200:9200"
-      - "9300:9300"
+  search:
+    build:
+      context: search
     environment:
       - node.name=es-node01
       - cluster.name=es-cluster
       - discovery.type=single-node
-      - "ES_JAVA_OPTS=-Xms128m -Xmx128m"
+      - "ES_JAVA_OPTS=-Xms512m -Xmx512m"
     ulimits:
       memlock:
         soft: -1
         hard: -1
+      nofile:
+        soft: 65536
+        hard: 65536
     volumes:
       - elasticsearch_data:/usr/share/elastic-setup/data
