+++ search/README.md
@@ -1,0 +1,11 @@
+### Incremental indexing
+
+```
+npm run index
+```
+
+### Full re-index
+
+```
+npm run re-index
+```
